+++ rv/project.py
@@ -35,6 +35,7 @@ class Project(Container):
         self.global_volume = 80
         self.name = 'Project'
         self.time_grid = 4
+        self.time_grid2 = 4
         self.metamodule = None
         self.modules_scale = 256
         self.modules_zoom = 256
@@ -104,6 +105,7 @@ class Project(Container):
         yield (b'BPM ', pack('<I', self.initial_bpm))
         yield (b'SPED', pack('<I', self.initial_tpl))
         yield (b'TGRD', pack('<I', self.time_grid))
+        yield (b'TGD2', pack('<I', self.time_grid2))
         yield (b'GVOL', pack('<I', self.global_volume))
         yield (b'NAME', self.name.encode(ENCODING) + b'\0')
         yield (b'MSCL', pack('<I', self.modules_scale))